refactor(core): Moving worker loader to a folder
